requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
antlr4-python3-runtime==4.8
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
black==22.1.0
bleach==4.1.0
blinker==1.4
braceexpand==0.1.7
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
clip==1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dlib==19.23.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==0.8.2
filelock==3.6.0
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.1.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.2.1
idna==3.3
imageio==2.9.0
importlib-metadata==4.11.1
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.4.1
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
omegaconf==2.1.1
open-clip-torch==0.2.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
pyDeprecate==0.3.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
pysift==0.0.2
python-dateutil==2.8.2
pytorch-lightning==1.5.2
pytorch-metric-learning==1.1.0
pytorch-pretrained-vit==0.0.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentry-sdk==1.5.5
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
streamlit==1.5.1
-e git+https://github.com/CompVis/taming-transformers@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.1+cu102
torch-fidelity==0.3.0
torchaudio==0.10.1+rocm4.1
torchmetrics==0.7.1
torchvision==0.11.2+cu102
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.18.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
wandb==0.12.10
watchdog==2.1.6
wcwidth==0.2.5
webdataset==0.1.103
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
einops